==> scripts/Anton_Victor_Student_B/00_B_DataScraping.ipynb <==
## Data Scraping from Markets Business Insider

from bs4 import BeautifulSoup
import requests
import pandas as pd

def scrape_index_components_to_csv(url, csv_filename):
    response = requests.get(url)
    soup = BeautifulSoup(response.text, 'html.parser')

    # Create a list to hold all the company data
    companies_data = []

    # Extract the header names for columns
    headers = soup.select_one('table.table thead').find_all('th')
    columns = [header.get_text(strip=True) for header in headers]

    # Loop through the rows and extract data
    for row in soup.select('table.table tbody tr'):
        # Get all columns in the row
        cols = row.find_all('td')
        # Extract text and split on new lines or other appropriate separators
        col_text = [col.get_text(" ", strip=True) for col in cols]
        companies_data.append(col_text)

    # Create the DataFrame
    df = pd.DataFrame(companies_data, columns=columns)

    # Clean the DataFrame to remove rows with no company name
    df = df.dropna(subset=[columns[0]])  # assuming the first column is the company name

    # Save the DataFrame as CSV
    df.to_csv(f'../data/raw_data/{csv_filename}', index=False)

    return df


def scrape_wikipedia_to_csv(url, csv_filename, table_position):
    response = requests.get(url)
    soup = BeautifulSoup(response.text, 'html.parser')

    # Since the tables have a common class, we find all and then select the third one
    tables = soup.find_all('table', {'class': 'wikitable sortable'})
    
    # Table position
    table = tables[table_position] 
    
    # Create a list to hold all the company data
    companies_data = []

    # Extract the header names for columns
    headers = table.find_all('th')
    columns = [header.get_text(strip=True) for header in headers]

    # Loop through the rows and extract data
    for row in table.find_all('tr'):
        cols = row.find_all('td')
        if cols:  # This checks if there are any 'td' elements found to avoid header rows
            col_text = [col.get_text(" ", strip=True) for col in cols]
            companies_data.append(col_text)

    # Create the DataFrame
    df = pd.DataFrame(companies_data, columns=columns)

    # Clean the DataFrame to remove rows with no company name
    # assuming the first column is the company name
    df = df.dropna(subset=[columns[0]])  

    # Save the DataFrame as CSV
    df.to_csv(f'../data/raw_data/{csv_filename}', index=False)


    return df

### AEX. Amsterdam Exchange Index

url = "https://markets.businessinsider.com/index/components/aex"
csv_filename = 'aex_stock_data.csv'
df = scrape_index_components_to_csv(url, csv_filename)
print(df)

url = "https://en.wikipedia.org/wiki/AEX_index"
csv_filename = 'aex_wikipedia_data.csv'
df = scrape_wikipedia_to_csv(url, csv_filename, 1)
print(df)

### BEL 20. Brussels Stock Exchange

url = "https://markets.businessinsider.com/index/components/bel_20"
csv_filename = 'bel_20_stock_data.csv'
df = scrape_index_components_to_csv(url, csv_filename)
print(df)

url = "https://en.wikipedia.org/wiki/BEL_20"
csv_filename = 'bel_20_wikipedia_data.csv'
df = scrape_wikipedia_to_csv(url, csv_filename, 1)
print(df)

### CAC 40. Cotation Assistée en Continu
Benchmark French Stock Market Index

url = "https://markets.businessinsider.com/index/components/cac_40"
csv_filename = 'cac_40_stock_data.csv'
df = scrape_index_components_to_csv(url, csv_filename)
print(df)

url = "https://en.wikipedia.org/wiki/CAC_40"
csv_filename = 'cac_40_wikipedia_data.csv'
df = scrape_wikipedia_to_csv(url, csv_filename, 1)
print(df)

### ISEQ 20. Ireland Overall Stock Exchange Index

url = "https://markets.businessinsider.com/index/components/iseq"
csv_filename = 'iseq_20_stock_data.csv'
df = scrape_index_components_to_csv(url, csv_filename)
print(df)

url = "https://en.wikipedia.org/wiki/ISEQ_20"
csv_filename = 'iseq_20_wikipedia_data.csv'
df = scrape_wikipedia_to_csv(url, csv_filename, 0)
print(df)

### OBX. Oslo Stock Exchange

url = "https://markets.businessinsider.com/index/components/obx"
csv_filename = 'obx_stock_data.csv'
df = scrape_index_components_to_csv(url, csv_filename)
print(df)

url = "https://en.wikipedia.org/wiki/OBX_Index"
csv_filename = 'obx_wikipedia_data.csv'
df = scrape_wikipedia_to_csv(url, csv_filename, 0)
print(df)

### OSEBX. Oslo Børs All Share Index

url = "https://markets.businessinsider.com/index/components/osebx"
csv_filename = 'osebx_stock_data.csv'
df = scrape_index_components_to_csv(url, csv_filename)
print(df)

url = "https://en.wikipedia.org/wiki/OSEAX"
csv_filename = 'osebx_wikipedia_data.csv'
df = scrape_wikipedia_to_csv(url, csv_filename, 1)
print(df)

### PSI. Portuguese Stock Index

url = "https://markets.businessinsider.com/index/components/psi20"
csv_filename = 'psi_20_stock_data.csv'
df = scrape_index_components_to_csv(url, csv_filename)
print(df)

url = "https://en.wikipedia.org/wiki/PSI-20"
csv_filename = 'psi_20_wikipedia_data.csv'
df = scrape_wikipedia_to_csv(url, csv_filename, 1)
print(df)
