==> scripts/Anton_Victor_Student_B/02_B_DataLoad_MariaDB.ipynb <==
import mariaDB
import pandas as pd
import sys


# Connect to MariaDB Platform
try:
    conn = mariaDB.connect(
        user="cip_user",
        password="cip_pw",
        host="127.0.0.1",                    # 10.177.124.185,  localhost , 127.0.0.1     
        #host="10.177.124.35",                    # 10.177.124.35,  localhost , 127.0.0.1    FS23
        port=3306, 
        database="CIP"

    )
except mariaDB.Error as e:
    print(f"Error connecting to MariaDB Platform: {e}")
    sys.exit(1)

# Get Cursor
cur = conn.cursor()

df = pd.read_csv('../data/clean_data/combined_final_data.csv')

# Iterate over each row in the DataFrame
for index, row in df.iterrows():
    # Extract data from the DataFrame row
    idx = row['Index']
    company = row['Company']
    ticker = row['Ticker']
    
    # Execute INSERT query
    try:
        cur.execute("INSERT INTO your_table_name (Index, Company, Ticker) VALUES (?, ?, ?)", (idx, company, ticker))
    except mariadb.Error as e:
        print(f"Error: {e}")
        continue

# Commit changes to the database
conn.commit()

# Print Result-set
cur.execute("SELECT Index, Company, Ticker FROM your_table_name")
for (idx, company, ticker) in cur:
    print(f"Index: {idx}, Company: {company}, Ticker: {ticker}")

# Close cursor and connection
cur.close()
conn.close()
